I changed floatX to be an int. The intention is to make all the relevant changes to this branch to get the framework to work with the int datatype

+++ nanolib2.0/tensor/tensor_util.hpp
@@ -5,7 +5,7 @@
 #include "tensor_types.hpp"
 #include "eigen/Eigen/Core"
 
-using floatX = float;
+using floatX = int;
 
 // Raw pointer -> Flat
 template <typename T>
